apps/api/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
asyncpg==0.30.0
psycopg2-binary==2.9.10
alembic==1.14.1
redis==5.2.1
celery==5.4.0
python-multipart==0.0.19
Pillow==11.1.0
pydantic==2.10.4
pydantic-settings==2.7.1
